Refactored the story card components so that now instead of one file containing multiple components, there are two distinct files each only housing one component. Those components use if statements to determine how the card should be displayed.
